+++ rust/operator-binary/src/kerberos.rs
@@ -1,21 +1,37 @@
 use indoc::formatdoc;
 use snafu::{ResultExt, Snafu};
 use stackable_hive_crd::{HiveCluster, HiveRole, HIVE_SITE_XML, STACKABLE_CONFIG_DIR};
-use stackable_operator::builder::pod::{
-    container::ContainerBuilder,
-    volume::{
-        SecretOperatorVolumeSourceBuilder, SecretOperatorVolumeSourceBuilderError, VolumeBuilder,
+use stackable_operator::{
+    builder::{
+        self,
+        pod::{
+            container::ContainerBuilder,
+            volume::{
+                SecretOperatorVolumeSourceBuilder, SecretOperatorVolumeSourceBuilderError,
+                VolumeBuilder,
+            },
+            PodBuilder,
+        },
     },
-    PodBuilder,
+    kube::ResourceExt,
+    utils::cluster_info::KubernetesClusterInfo,
 };
-use stackable_operator::kube::ResourceExt;
 use std::collections::BTreeMap;
 
 #[derive(Snafu, Debug)]
+#[allow(clippy::enum_variant_names)] // all variants have the same prefix: `Add`
 pub enum Error {
     #[snafu(display("failed to add Kerberos secret volume"))]
     AddKerberosSecretVolume {
         source: SecretOperatorVolumeSourceBuilderError,
+    },
+
+    #[snafu(display("failed to add needed volume"))]
+    AddVolume { source: builder::pod::Error },
+
+    #[snafu(display("failed to add needed volumeMount"))]
+    AddVolumeMount {
+        source: builder::pod::container::Error,
     },
 }
 
@@ -37,8 +53,10 @@ pub fn add_kerberos_pod_config(
             VolumeBuilder::new("kerberos")
                 .ephemeral(kerberos_secret_operator_volume)
                 .build(),
-        );
-        cb.add_volume_mount("kerberos", "/stackable/kerberos");
+        )
+        .context(AddVolumeSnafu)?;
+        cb.add_volume_mount("kerberos", "/stackable/kerberos")
+            .context(AddVolumeMountSnafu)?;
 
         // Needed env vars
         cb.add_env_var("KRB5_CONFIG", "/stackable/kerberos/krb5.conf");
@@ -50,13 +68,16 @@ pub fn add_kerberos_pod_config(
 pub fn kerberos_config_properties(
     hive: &HiveCluster,
     hive_namespace: &str,
+    cluster_info: &KubernetesClusterInfo,
 ) -> BTreeMap<String, String> {
     if !hive.has_kerberos_enabled() {
         return BTreeMap::new();
     }
+
     let hive_name = hive.name_any();
+    let cluster_domain = &cluster_info.cluster_domain;
     let principal_host_part =
-        format!("{hive_name}.{hive_namespace}.svc.cluster.local@${{env.KERBEROS_REALM}}");
+        format!("{hive_name}.{hive_namespace}.svc.{cluster_domain}@${{env.KERBEROS_REALM}}");
 
     BTreeMap::from([
         // Kerberos settings