+++ rust/operator-binary/src/main.rs
@@ -17,6 +17,7 @@ use stackable_operator::{
         apps::v1::StatefulSet,
         core::v1::{ConfigMap, Service},
     },
+    kube::core::DeserializeGuard,
     kube::runtime::{watcher, Controller},
     logging::controller::report_controller_reconciled,
     CustomResourceExt,
@@ -45,6 +46,7 @@ async fn main() -> anyhow::Result<()> {
             product_config,
             watch_namespace,
             tracing_target,
+            cluster_info_opts,
         }) => {
             stackable_operator::logging::initialize_logging(
                 "HIVE_OPERATOR_LOG",
@@ -65,11 +67,14 @@ async fn main() -> anyhow::Result<()> {
                 "/etc/stackable/hive-operator/config-spec/properties.yaml",
             ])?;
 
-            let client =
-                stackable_operator::client::create_client(Some(OPERATOR_NAME.to_string())).await?;
+            let client = stackable_operator::client::initialize_operator(
+                Some(OPERATOR_NAME.to_string()),
+                &cluster_info_opts,
+            )
+            .await?;
 
             Controller::new(
-                watch_namespace.get_api::<HiveCluster>(&client),
+                watch_namespace.get_api::<DeserializeGuard<HiveCluster>>(&client),
                 watcher::Config::default(),
             )
             .owns(